Adapted model
As soon as I'll get it going on the uni computer I'll add requirements file. I'm still working on data processing so it is well maintained. I'll add it till Friday. Also I had to skip one training file because it was to big for github. I'll add a link to the original repo so it can be dwonloaded (probably github restricted file sizes which can be stored in the project over the years)